feat: Split aot mode into its own option

+++ test.sh
@@ -12,6 +12,16 @@ PATH=$PATH:$RISCV/bin
 TOP="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 cd $TOP
 
+if [ "$1" = "--enable-aot" ]
+then
+    BUILD_OPTIONS="--features=aot"
+    AOT=1
+    shift
+else
+    BUILD_OPTIONS=""
+    AOT=0
+fi
+
 # If requested, make sure we are using latest revision of CKB VM
 if [ "$1" = "--update-ckb-vm" ]
 then
@@ -29,22 +39,30 @@ then
     INTERPRETER32="kcov --verify $TOP/coverage $TOP/binary/target/debug/interpreter32"
     INTERPRETER64="kcov --verify $TOP/coverage $TOP/binary/target/debug/interpreter64"
     ASM64="kcov --verify $TOP/coverage $TOP/binary/target/debug/asm64"
-    AOT64="kcov --verify $TOP/coverage $TOP/binary/target/debug/aot64"
+
+    if [ "$AOT" -eq "1" ]
+    then
+        AOT64="kcov --verify $TOP/coverage $TOP/binary/target/debug/aot64"
+    fi
 
     rm -rf $TOP/coverage
 
     # Build CKB VM binaries for testing
     cd "$TOP/binary"
-    cargo build
+    cargo build $BUILD_OPTIONS
 else
     INTERPRETER32="$TOP/binary/target/release/interpreter32"
     INTERPRETER64="$TOP/binary/target/release/interpreter64"
     ASM64="$TOP/binary/target/release/asm64"
-    AOT64="$TOP/binary/target/release/aot64"
+
+    if [ "$AOT" -eq "1" ]
+    then
+        AOT64="$TOP/binary/target/release/aot64"
+    fi
 
     # Build CKB VM binaries for testing
     cd "$TOP/binary"
-    cargo build --release
+    cargo build --release $BUILD_OPTIONS
 fi
 
 
@@ -67,9 +85,13 @@ done
 for i in $(find . -regex ".*/rv64u[imc]-u-[a-z0-9_]*" | grep -v "fence_i" | grep -v "rv64ui-u-jalr"); do
     $ASM64 $i
 done
-for i in $(find . -regex ".*/rv64u[imc]-u-[a-z0-9_]*" | grep -v "fence_i" | grep -v "rv64ui-u-jalr"); do
-    $AOT64 $i
-done
+
+if [ "$AOT" -eq "1" ]
+then
+    for i in $(find . -regex ".*/rv64u[imc]-u-[a-z0-9_]*" | grep -v "fence_i" | grep -v "rv64ui-u-jalr"); do
+        $AOT64 $i
+    done
+fi
 
 # Test CKB VM with riscv-compliance
 cd "$TOP/riscv-compliance"
@@ -81,9 +103,13 @@ rm -rf work && make RISCV_TARGET=ckb-vm XLEN=64 RISCV_DEVICE=C TARGET_SIM="$INTE
 rm -rf work && make RISCV_TARGET=ckb-vm XLEN=64 RISCV_DEVICE=I TARGET_SIM="$ASM64" simulate
 rm -rf work && make RISCV_TARGET=ckb-vm XLEN=64 RISCV_DEVICE=M TARGET_SIM="$ASM64" simulate
 rm -rf work && make RISCV_TARGET=ckb-vm XLEN=64 RISCV_DEVICE=C TARGET_SIM="$ASM64" simulate
-rm -rf work && make RISCV_TARGET=ckb-vm XLEN=64 RISCV_DEVICE=I TARGET_SIM="$AOT64" simulate
-rm -rf work && make RISCV_TARGET=ckb-vm XLEN=64 RISCV_DEVICE=M TARGET_SIM="$AOT64" simulate
-rm -rf work && make RISCV_TARGET=ckb-vm XLEN=64 RISCV_DEVICE=C TARGET_SIM="$AOT64" simulate
+
+if [ "$AOT" -eq "1" ]
+then
+    rm -rf work && make RISCV_TARGET=ckb-vm XLEN=64 RISCV_DEVICE=I TARGET_SIM="$AOT64" simulate
+    rm -rf work && make RISCV_TARGET=ckb-vm XLEN=64 RISCV_DEVICE=M TARGET_SIM="$AOT64" simulate
+    rm -rf work && make RISCV_TARGET=ckb-vm XLEN=64 RISCV_DEVICE=C TARGET_SIM="$AOT64" simulate
+fi
 
 # Even though ckb-vm-bench-scripts are mainly used for benchmarks, they also
 # contains sophisticated scripts which make good tests
@@ -91,10 +117,14 @@ cd "$TOP/ckb-vm-bench-scripts"
 make
 $INTERPRETER64 build/secp256k1_bench 033f8cf9c4d51a33206a6c1c6b27d2cc5129daa19dbd1fc148d395284f6b26411f 304402203679d909f43f073c7c1dcf8468a485090589079ee834e6eed92fea9b09b06a2402201e46f1075afa18f306715e7db87493e7b7e779569aa13c64ab3d09980b3560a3 foo bar
 $ASM64 build/secp256k1_bench 033f8cf9c4d51a33206a6c1c6b27d2cc5129daa19dbd1fc148d395284f6b26411f 304402203679d909f43f073c7c1dcf8468a485090589079ee834e6eed92fea9b09b06a2402201e46f1075afa18f306715e7db87493e7b7e779569aa13c64ab3d09980b3560a3 foo bar
-$AOT64 build/secp256k1_bench 033f8cf9c4d51a33206a6c1c6b27d2cc5129daa19dbd1fc148d395284f6b26411f 304402203679d909f43f073c7c1dcf8468a485090589079ee834e6eed92fea9b09b06a2402201e46f1075afa18f306715e7db87493e7b7e779569aa13c64ab3d09980b3560a3 foo bar
 
 $INTERPRETER64 build/schnorr_bench 4103c5b538d6f695a961e916e7308211c8c917e1e02ca28a21b0989596a9ffb6 e45408b5981ec7fd6e72faa161776fe5db17dd92226d1ad784816fb843e151127d9ccb615f364f317a35e2ddddc91bbf30ad103ddfd3ad7e839f508dbfe6298a foo bar
 $ASM64 build/schnorr_bench 4103c5b538d6f695a961e916e7308211c8c917e1e02ca28a21b0989596a9ffb6 e45408b5981ec7fd6e72faa161776fe5db17dd92226d1ad784816fb843e151127d9ccb615f364f317a35e2ddddc91bbf30ad103ddfd3ad7e839f508dbfe6298a foo bar
-$AOT64 build/schnorr_bench 4103c5b538d6f695a961e916e7308211c8c917e1e02ca28a21b0989596a9ffb6 e45408b5981ec7fd6e72faa161776fe5db17dd92226d1ad784816fb843e151127d9ccb615f364f317a35e2ddddc91bbf30ad103ddfd3ad7e839f508dbfe6298a foo bar
+
+if [ "$AOT" -eq "1" ]
+then
+    $AOT64 build/secp256k1_bench 033f8cf9c4d51a33206a6c1c6b27d2cc5129daa19dbd1fc148d395284f6b26411f 304402203679d909f43f073c7c1dcf8468a485090589079ee834e6eed92fea9b09b06a2402201e46f1075afa18f306715e7db87493e7b7e779569aa13c64ab3d09980b3560a3 foo bar
+    $AOT64 build/schnorr_bench 4103c5b538d6f695a961e916e7308211c8c917e1e02ca28a21b0989596a9ffb6 e45408b5981ec7fd6e72faa161776fe5db17dd92226d1ad784816fb843e151127d9ccb615f364f317a35e2ddddc91bbf30ad103ddfd3ad7e839f508dbfe6298a foo bar
+fi
 
 echo "All tests are passed!"
